issue/#29 fix : 認証機能に伴うjsxファイルの追加など
