+++ package.json
@@ -8,13 +8,16 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "axios": "^1.3.4",
+    "axios": "^1.3.5",
+    "axios-case-converter": "^1.1.0",
     "bootstrap": "^5.2.3",
     "dayjs": "^1.11.7",
+    "js-cookie": "^3.0.1",
     "json-server": "^0.17.2",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.2",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "5.0.1",
     "recoil": "^0.7.6",
     "sass": "^1.58.3",
