+++ src/App.jsx
@@ -1,0 +1,78 @@
+import { createContext, useEffect, useState } from "react";
+import { BrowserRouter, Switch, Route, Redirect } from "react-router-dom";
+import { getCurrentUser } from "./api/auth";
+import SignIn from "./auth/components/SignIn";
+import SignUp from "./auth/components/SignUp";
+import Home from "./home/screen"
+
+export const AuthContext = createContext("");
+
+function App() {
+  const [loading, setLoading] = useState(true);
+  const [isSignedIn, setIsSignedIn] = useState(false);
+  const [currentUser, setCurrentUser] = useState();
+
+  const handleGetCurrentUser = async () => {
+    try {
+      const res = await getCurrentUser();
+
+      if (res?.data.isLogin === true) {
+        setIsSignedIn(true);
+        setCurrentUser(res?.data.data);
+        console.log(res?.data.data);
+      } else {
+        console.log("no current user");
+      }
+    } catch (e) {
+      console.log(e);
+    }
+    setLoading(false);
+  };
+
+  useEffect(() => {
+    handleGetCurrentUser();
+  }, [setCurrentUser]);
+
+  const Private = ({ children }) => {
+    if (!loading) {
+      if (isSignedIn) {
+        return children;
+      } else {
+        return <Redirect to="signin" />;
+      }
+    } else {
+      return <></>;
+    }
+  };
+  
+  return (
+    <AuthContext.Provider
+      value={{
+        loading,
+        setLoading,
+        isSignedIn,
+        setIsSignedIn,
+        currentUser,
+        setCurrentUser,
+      }}
+    >
+      <BrowserRouter>
+        <Switch>
+          <Route exact path="/signup">
+            <SignUp />
+          </Route>
+          <Route exact path="/signin">
+            <SignIn />
+          </Route>
+          <Private>
+            <Route exact path="/">
+              <Home />
+            </Route>
+          </Private>
+        </Switch>
+      </BrowserRouter>
+    </AuthContext.Provider>
+  );
+}
+
+export default App;
