+++ src/auth/screen/index.jsx
@@ -1,76 +1,76 @@
-import { createContext, useEffect, useState } from "react";
-import { BrowserRouter, Switch, Route, Redirect } from "react-router-dom";
-import { getCurrentUser } from "../api/auth";
-import { SignIn } from "../components/SignIn";
-import { SignUp } from "../components/SignUp";
+// import { createContext, useEffect, useState } from "react";
+// import { BrowserRouter, Switch, Route, Redirect } from "react-router-dom";
+// import { getCurrentUser } from "../api/auth";
+// import { SignIn } from "../components/SignIn";
+// import { SignUp } from "../components/SignUp";
 
-export const AuthContext = createContext();
+// export const AuthContext = createContext();
 
-function App() {
-  const [loading, setLoading] = useState(true);
-  const [isSignedIn, setIsSignedIn] = useState(false);
-  const [currentUser, setCurrentUser] = useState();
+// function App() {
+//   const [loading, setLoading] = useState(true);
+//   const [isSignedIn, setIsSignedIn] = useState(false);
+//   const [currentUser, setCurrentUser] = useState();
 
-  const handleGetCurrentUser = async () => {
-    try {
-      const res = await getCurrentUser();
+//   const handleGetCurrentUser = async () => {
+//     try {
+//       const res = await getCurrentUser();
 
-      if (res?.data.isLogin === true) {
-        setIsSignedIn(true);
-        setCurrentUser(res?.data.data);
-        console.log(res?.data.data);
-      } else {
-        console.log("no current user");
-      }
-    } catch (e) {
-      console.log(e);
-    }
-    setLoading(false);
-  };
+//       if (res?.data.isLogin === true) {
+//         setIsSignedIn(true);
+//         setCurrentUser(res?.data.data);
+//         console.log(res?.data.data);
+//       } else {
+//         console.log("no current user");
+//       }
+//     } catch (e) {
+//       console.log(e);
+//     }
+//     setLoading(false);
+//   };
 
-  useEffect(() => {
-    handleGetCurrentUser();
-  }, [setCurrentUser]);
+//   useEffect(() => {
+//     handleGetCurrentUser();
+//   }, [setCurrentUser]);
 
-  const Private = ({ children }) => {
-    if (!loading) {
-      if (isSignedIn) {
-        return children;
-      } else {
-        return <Redirect to="signin" />;
-      }
-    } else {
-      return <></>;
-    }
-  };
-  return (
-    <AuthContext.Provider
-      value={{
-        loading,
-        setLoading,
-        isSignedIn,
-        setIsSignedIn,
-        currentUser,
-        setCurrentUser,
-      }}
-    >
-      <BrowserRouter>
-        <Switch>
-          <Route exact path="/signup">
-            <SignUp />
-          </Route>
-          <Route exact path="/signin">
-            <SignIn />
-          </Route>
-          <Private>
-            <Route exact path="/">
-              <Home />
-            </Route>
-          </Private>
-        </Switch>
-      </BrowserRouter>
-    </AuthContext.Provider>
-  );
-}
+//   const Private = ({ children }) => {
+//     if (!loading) {
+//       if (isSignedIn) {
+//         return children;
+//       } else {
+//         return <Redirect to="signin" />;
+//       }
+//     } else {
+//       return <></>;
+//     }
+//   };
+//   return (
+//     <AuthContext.Provider
+//       value={{
+//         loading,
+//         setLoading,
+//         isSignedIn,
+//         setIsSignedIn,
+//         currentUser,
+//         setCurrentUser,
+//       }}
+//     >
+//       <BrowserRouter>
+//         <Switch>
+//           <Route exact path="/signup">
+//             <SignUp />
+//           </Route>
+//           <Route exact path="/signin">
+//             <SignIn />
+//           </Route>
+//           <Private>
+//             <Route exact path="/">
+//               <Home />
+//             </Route>
+//           </Private>
+//         </Switch>
+//       </BrowserRouter>
+//     </AuthContext.Provider>
+//   );
+// }
 
-export default App;
+// export default App;
