+++ src/core/components/Router/index.js
@@ -1,5 +1,7 @@
 import { ReactLocation } from '@tanstack/react-location';
 import Home from '../../../home/screen';
+import SignIn from '../../../auth/components/SignIn'
+import SignUp from '../../../auth/components/SignUp'
 import RecoilButton from '../../../recoil/Button';
 
 export const location = new ReactLocation();
@@ -8,6 +10,14 @@ export const routes = [
   {
     path: '/',
     element: <Home />,
+  },
+  {
+    path: '/signup',
+    element: <SignUp />
+  },
+  {
+    path: '/signin',
+    element: <SignIn />
   },
   {
     path: 'recoil',
